Optimized I/O
When there are multiple input operations without any outputs inbetween, or vice versa,
we don't need to set the stream for every operation. Sadly, we can't optimize across
loop boundaries, because as it turns out, an optimizer for that would effectively be
a halting problem solver. This is what we know in the industry as a fat bruh moment.